refactor(cli): Treat -Q as an array and removed the custom parsing based on a comma.
Until now the `-Q` argument accepted a comma separated string to supply multiple
queries. This made it impossible to supply more complx query terms, such as JSON
strings. This behavior is deprecated from now, but instead supply `-Q` as an
array of arguments. Both of the following approaches are valid:

    sugarcube -Q query_type:term1 -Q query_type:term2
    sugarcube -Q query_type:term1 query_type:term2