refactor(plugin-googlesheets): Use the builtin cache facility for storing the auth credentials.
Until now all Google sheet plugins stored credentials in a custom file called
`google-sheets-token.json`. This mechanism is deprecated now and credentials
are stored in the SugarCube cache instead. For the CLI runner this means a file
named `.sugarcube_cache.json` in the project root.

The simplest way to upgrade is to run a pipeline and authorize the app again
with Google.