fix: Handle media urls better to avoid redownloads.
If any media URL contains a changing query parameter, the `_sc_id_hash` is
different for every run. This triggers unnecessary downloads and duplicated
media downloads. This shows particularly for facebook image and video media.

If the media download includes a separate `href` field, use that over term
as the download target URL. `term` can then be a more unique element of the
url, such as the filename.

The facebook plugins make use of this. The http_get, http_wget,
media_youtubedl and media_screenshot plugins detect the new href field.

Fixes #33